+++ src/libmongoc/src/mongoc/mongoc-database.c
@@ -30,7 +30,7 @@
 #include "mongoc-write-concern-private.h"
 #include "mongoc-change-stream-private.h"
 
-#include <bson/bson-dsl.h>
+#include <bson-dsl.h>
 
 #undef MONGOC_LOG_DOMAIN
 #define MONGOC_LOG_DOMAIN "database"
