+++ src/libmongoc/tests/unified/entity-map.c
@@ -23,7 +23,7 @@
 #include "utlist.h"
 #include "util.h"
 
-#include <bson/bson-dsl.h>
+#include <bson-dsl.h>
 
 #include "common-b64-private.h"
 #include "mongoc-client-side-encryption-private.h"
